Completed Authorization For Place Only Creator Of The PLace Can Remove And Update A Place

+++ src/Backend/Place/Controller/PlaceController.js
@@ -5,10 +5,11 @@ const AppError = require("../../Utils/AppError");
 const fs = require("fs");
 
 exports.CreatePlace = AsyncWrapper(async (req, res, next) => {
-  const { title, description, address, creator } = req.body;
+  const { title, description, address, creator, image } = req.body;
   const Place = new PlaceModel({
     title,
     description,
+    image,
     image: req.file.path,
     address,
     creator
@@ -73,13 +74,18 @@ exports.UpdatePlace = AsyncWrapper(async (req, res, next) => {
     return next(new AppError("Unable To Find The Place With Provided Id", 404));
   }
   const { title, description } = req.body;
+
   Place.title = title;
   Place.description = description;
-  await Place.save();
-  res.status(200).json({
-    Status: "Success",
-    Place
-  });
+  if (req.userId === Place.creator._id.toString()) {
+    await Place.save();
+    res.status(200).json({
+      Status: "Success",
+      Place
+    });
+  } else {
+    return next(new AppError("You Are Not Authorized For This Action", 401));
+  }
 });
 
 exports.DeletePlace = AsyncWrapper(async (req, res, next) => {
@@ -88,13 +94,17 @@ exports.DeletePlace = AsyncWrapper(async (req, res, next) => {
   if (!Place) {
     return next(new AppError("Unable To Find The Place With Provided Id", 404));
   }
-  await Place.remove();
-  Place.creator.places.pull(Place);
-  await Place.creator.save();
-  // It Will Delete The Picure Of Image From Static Folder-
-  fs.unlink(Place.image, err => {});
-  res.status(200).json({
-    Status: "Sucess",
-    Message: "Deleted Item With Id " + Id + " Has Been Deleted :)"
-  });
+  if (req.userId === Place.creator._id.toString()) {
+    await Place.remove();
+    Place.creator.places.pull(Place);
+    await Place.creator.save();
+    // It Will Delete The Picure Of Image From Static Folder-
+    fs.unlink(Place.image, err => {});
+    res.status(200).json({
+      Status: "Sucess",
+      Message: "Deleted Item With Id " + Id + " Has Been Deleted :)"
+    });
+  } else {
+    return next(new AppError("You Are Not Authorized For This Action", 401));
+  }
 });
